# Backend/api-docs/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8443
  description: Generated server url
paths:
...
  /api/v1/users/{id}/teams/{teamId}:
    put:
      tags:
      - user-controller
      summary: Update an User matches
      operationId: addUserTeams
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: teamId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: Bad Request
        "422":
          description: Unprocessable Entity
        "401":
          description: Unauthorized
        "404":
          description: User not found
        "200":
          description: User teams Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamResponseDTO"
    delete:
      tags:
      - user-controller
      summary: Delete an User team
      operationId: deleteUserTeams
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: teamId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "401":
          description: Unauthorized
        "404":
          description: User not found
        "200":
          description: User team Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamResponseDTO"
...
  /api/v1/teams/{id}:
    get:
      tags:
      - team-controller
      summary: Get team by id
      operationId: getTeam
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: team not found
        "200":
          description: Found team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamResponseDTO"
    put:
      tags:
      - team-controller
      summary: Update a Team
      operationId: putTeams
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamUpdateDTO"
        required: true
      responses:
        "404":
          description: Team not found
        "200":
          description: Team Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamResponseDTO"
        "400":
          description: Bad Request
        "422":
          description: Unprocessable Entity
        "401":
          description: Unauthorized
    delete:
      tags:
      - team-controller
      summary: Delete a Team
      operationId: deleteTeams
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Team not found
        "200":
          description: Team Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "401":
          description: Unauthorized
...
  /api/v1/users/{id}/image:
    get:
      tags:
      - user-controller
      summary: Get user image
      operationId: getImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: user not found
        "200":
          description: Found user image
          content:
            application/json:
              schema:
                type: object
                properties:
                  binaryStream:
                    type: object
    post:
      tags:
      - user-controller
      summary: Create an User image
      operationId: uploadImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              required:
              - imageFile
              type: object
              properties:
                imageFile:
                  type: string
                  format: binary
      responses:
        "201":
          description: User Image Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponseDTO"
        "400":
          description: Bad Request
        "422":
          description: Unprocessable Entity
        "401":
          description: Unauthorized
        "409":
          description: Conflict
        "403":
          description: Forbidden
    delete:
      tags:
      - user-controller
      summary: Delete an User image
      operationId: deleteImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: User Image Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponseDTO"
        "401":
          description: Unauthorized
        "404":
          description: User not found
        "403":
          description: Forbidden
...
  /api/v1/teams/{id}/image:
    get:
      tags:
      - team-controller
      summary: Get team image
      operationId: getImage_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: team not found
        "200":
          description: Found team image
          content:
            application/json:
              schema:
                type: object
                properties:
                  binaryStream:
                    type: object
    post:
      tags:
      - team-controller
      summary: Create a Team image
      operationId: uploadImage_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              required:
              - imageFile
              type: object
              properties:
                imageFile:
                  type: string
                  format: binary
      responses:
        "201":
          description: Team Image Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamResponseDTO"
        "400":
          description: Bad Request
        "422":
          description: Unprocessable Entity
        "401":
          description: Unauthorized
        "409":
          description: Conflict
        "403":
          description: Forbidden
    delete:
      tags:
      - team-controller
      summary: Delete a Team image
      operationId: deleteImage_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Team Image Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamResponseDTO"
        "404":
          description: Team not found
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /api/v1/teams/:
    get:
      tags:
      - team-controller
      summary: Get all the teams
      operationId: getAllTeams
      responses:
        "200":
          description: Found teams
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamResponseDTO"
        "204":
          description: No content
    post:
      tags:
      - team-controller
      summary: Create a team
      operationId: postTeams
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamCreationDTO"
        required: true
      responses:
        "400":
          description: Bad Request
        "201":
          description: Team Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamResponseDTO"
        "422":
          description: Unprocessable Entity
        "401":
          description: Unauthorized
        "409":
          description: Conflict
        "403":
          description: Forbidden
...
  /api/v1/players/{id}/image:
    get:
      tags:
      - playercontroller
      summary: Get player image
      operationId: getImage_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Found player image
          content:
            application/json:
              schema:
                type: object
                properties:
                  binaryStream:
                    type: object
        "404":
          description: player not found
    post:
      tags:
      - playercontroller
      summary: Create a League image
      operationId: uploadImage_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              required:
              - imageFile
              type: object
              properties:
                imageFile:
                  type: string
                  format: binary
      responses:
        "400":
          description: Bad Request
        "201":
          description: League Image Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerDTO"
        "422":
          description: Unprocessable Entity
        "401":
          description: Unauthorized
        "409":
          description: Conflict
        "403":
          description: Forbidden
    delete:
      tags:
      - playercontroller
      summary: Delete a Player image
      operationId: deleteImage_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: League not found
        "401":
          description: Unauthorized
        "200":
          description: Player Image Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerDTO"
        "403":
          description: Forbidden
...
  /api/v1/leagues/{id}/image:
    get:
      tags:
      - league-controller
      summary: Get league image
      operationId: getImage_3
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Image not found
        "200":
          description: Found image
          content:
            application/json:
              schema:
                type: object
                properties:
                  binaryStream:
                    type: object
    post:
      tags:
      - league-controller
      summary: Create a League image
      operationId: uploadImage_3
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              required:
              - imageFile
              type: object
              properties:
                imageFile:
                  type: string
                  format: binary
      responses:
        "400":
          description: Bad Request
        "422":
          description: Unprocessable Entity
        "201":
          description: League Image Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeagueDTO"
        "401":
          description: Unauthorized
        "409":
          description: Conflict
        "403":
          description: Forbidden
    delete:
      tags:
      - league-controller
      summary: Delete a League image
      operationId: deleteImage_3
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: League not found
        "200":
          description: League Image Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeagueDTO"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
...
  /api/v1/users/{id}/teams:
    get:
      tags:
      - user-controller
      summary: Get teams of a user
      operationId: getUserTeams
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: user not found
        "200":
          description: Found user teams
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamResponseDTO"
...
  /api/v1/teams/{teamsId}/matches:
    get:
      tags:
      - team-controller
      summary: Get matches of a team
      operationId: getMatchesByTeam
      parameters:
      - name: teamsId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: team not found
        "200":
          description: Found team matches
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchDTO"
  /api/v1/teams/{teamId}/players:
    get:
      tags:
      - team-controller
      summary: Get players of a team
      operationId: getPlayersByTeam
      parameters:
      - name: teamId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: team not found
        "200":
          description: Found team players
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerDTO"
  /api/v1/teams/{teamId}/league:
    get:
      tags:
      - team-controller
      summary: Get league of a team
      operationId: getLeagueByTeam
      parameters:
      - name: teamId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: team not found
        "200":
          description: Found team league
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeagueDTO"
  /api/v1/teams/name/{name}:
    get:
      tags:
      - team-controller
      summary: Get team by name
      operationId: getTeamByName
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      responses:
        "404":
          description: team not found
        "200":
          description: Found team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamResponseDTO"
  /api/v1/players/{playerId}/team:
    get:
      tags:
      - playercontroller
      summary: Get team of a player
      operationId: getTeamByPlayer
      parameters:
      - name: playerId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Found player team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamResponseDTO"
        "404":
          description: player not found
...
  /api/v1/matches/{matchId}/league:
    get:
      tags:
      - match-controller
      summary: Get league of a match
      operationId: getLeague
      parameters:
      - name: matchId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Found match league
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeagueDTO"
        "404":
          description: match not found
  /api/v1/matches/{id}/team2:
    get:
      tags:
      - match-controller
      summary: Get team2 of a match
      operationId: getTeam2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: match not found
        "200":
          description: Found match team2
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamResponseDTO"
  /api/v1/matches/{id}/team1:
    get:
      tags:
      - match-controller
      summary: Get team1 of a match
      operationId: getTeam1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: match not found
        "200":
          description: Found match team1
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamResponseDTO"
...
  /api/v1/leagues/{leagueId}/teams:
    get:
      tags:
      - league-controller
      summary: Get teams of a league
      operationId: getTeams
      parameters:
      - name: leagueId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Found league teams
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamResponseDTO"
        "204":
          description: No content
...
components:
  schemas:
...
    UserResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        roles:
          type: array
          items:
            type: string
        image:
          type: boolean
...
    PlayerDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        league:
          type: string
        team:
          type: string
        name:
          type: string
        age:
          type: integer
          format: int32
        nationality:
          type: string
        position:
          type: string
        image:
          type: boolean
    LeagueDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        president:
          type: string
        nationality:
          type: string
        image:
          type: boolean
    TeamUpdateDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        league:
          type: string
        trophies:
          type: integer
          format: int32
        nationality:
          type: string
        trainer:
          type: string
        secondTrainer:
          type: string
        president:
          type: string
        stadium:
          type: string
        image:
          type: boolean
...